--- app/actions/get-auth-session.ts
@@ -1,10 +1,0 @@
-import { getKindeServerSession } from '@kinde-oss/kinde-auth-nextjs/server';
-
-export function getAuthSession() {
-  const { getUser, isAuthenticated } = getKindeServerSession();
-
-  return {
-    user: getUser(),
-    isAuthenticated: isAuthenticated(),
-  };
-}
