+++ app/api/auth/[kindeAuth]/route.ts
